Add t.Parallel() to internal/vterm unit tests
The vterm suite used zero t.Parallel(). vterm is the ideal low-risk
starting point: each test builds an isolated New(w,h) terminal with no
package-level mutable state, no t.Setenv, and no shared resources. Add
t.Parallel() as the first statement of all 124 top-level test functions
(and the one table-driven subtest), skipping the two Fuzz entry points.

Adding t.Parallel() pushed prepend_scrollback_test.go over the 500-line
guard, so split its LoadPaneCapture group into
prepend_scrollback_load_test.go.
