+++ internal/vterm/pane_capture_alt_cursor_test.go
@@ -3,6 +3,7 @@ package vterm
 import "testing"
 
 func TestLoadPaneCaptureWithCursorAndModes_ClampsAltSavedCursorAfterResize(t *testing.T) {
+	t.Parallel()
 	vt := New(10, 3)
 	vt.Write([]byte("shell"))
 	vt.enterAltScreen()
